requirements.txt
streamlit>=1.30,<1.40
mysql-connector-python>=8.0
pandas>=2.0
